Holy shit stop listing my packages twice
What is wrong with me?  Why didn't I think of this sooner?  If
package.el is going to insist on keeping `package-selected-packages`
then just let that be my master list of packages.

I have the feeling I'm going to get ass-bit by this soon, but let's at
least give it a whirl.