Apparently multiple-cursors-core defines mc/keymap
I'm not sure if something changed in multiple-cursors, or in
Emacs (unlikely), or if I just had a broken Emacs at the time, but
multiple-cursors got loaded without multiple-cursors-core after a
recent update of both Emacs and packages.  As of this writing,
multiple-cursors.el doesn't seem to directly require
multiple-cursors-core, which is where mc/keymap is defined.