Fix persp-mode frameset saving
I didn't notice at exit that persp-mode was bitching about
un-printable values in the `frameset-save` value I was putting in a
property.  (How would I?  It logs it... then Emacs exits.)  Now we go
out of our way not to put anything un-printable (really un-readable)
in there.

Another bug fixed: need to put the property on the default/nil
perspective, just so we know where to find it to restore.  Otherwise
it'll get stashed in whatever the current perspective is at exit, but
then at startup it'll be in the nil perspective so you won't find
it (unless you were in nil perspective when you exited---which I was
in most of my testing, clearly).