sql-indent: use my fork, more fixes/configuration
All the el-patch-ing was getting too complicated for my tastes so I
just went ahead and forked the package and moved some of the patches
there, the ones I could come up with a test for.  Want to move
some/all of them to my fork in the future.

Other sql-indent-related changes here:

* No longer using `my:sqlind-indent-first-select-table-on-new-line`,
  in favor of `my:sqlind-indent-to-start-of-anchor-line` which gets me
  where I want to go, along with my new configuration.  Bonus: it's
  much simpler.

* `my:sqlind-indent-line-comment` was just broken, now fixed/improved.

* Fix probable bug in `sqlind-same-level-statement`.  This is one of
  those sql-indent patches I mentioned above for which I couldn't
  quickly come up with a test.  Damn me for not noting a test case as
  a comment when I wrote the patch.