Command to fetch a URL and stuff it in a buffer
`browse-url-emacs` should probably do the trick for me, but as of this
writing Emacs master has some kind of bug that makes it hork hard when
a buffer's file name is a URL, especially when you do **not** have
`url-handler-mode` on.  (That has its own problems with
`ivy-use-virtual-buffers` and recentf.)

This command also tries to guess a major mode in an awful, terribly
way.  See comments.