Unfuck persp-mode configuration
For the longest time I have been having Weird Shit™ happen when
persp-mode restores buffers/windows/frames.  Most recently it has been
local variables not getting applied.

Turns out YOU'RE REALLY NOT SUPPOSED TO TURN ON `persp-mode` IN YOUR
STARTUP.  You're supposed to turn it on in `after-init-hook`.  The
comments at the top of `persp-mode.el` even say so!  I feel **stupid**
for having missed this for so long.

I regrouped a bunch of the code here while I was at it.