Move stuff from my todo.org/todo.el into init.el
It was a PITA to have to use `eval` as a local variable, and I never
accepted it as a permanent safe value, so I was prompted every time at
startup.  Putting this Elisp into my main init.el is presumably a
better idea.  No more prompts at startup, now.