Rip out ivy-posframe
It overrides lots of default stuff like `ivy-avy` and the M-o binding,
and it does it in ways that assume you will be using ivy-posframe for
everything, which I definitely am I not.  I don't have time to fix
this right now, and I just don't care about it enough to keep it.
C-M-i goes back to the normal ol' minibuffer for me!