Fix a bunch of byte compilation warnings/errors
This is what I get for running

    emacs --batch --eval '(package-initialize)' -L $PWD \
        batch-byte-compile *.el

in my `lisp` and `recipes` directories.  Turns out I don't really
understand  byte compilation very well.  I probably only understand it
marginally better now.  Anyway, this was all born of the fact that I
ran the above and something that was a macro wasn't compiled as a
macro, so the whole thing (indirectly) blew up my `init.el`.
Hopefully the situation is a little better now, and as a bonus I might
be loading fewer things at startup.

`shackle-dismiss-pop-up-window.el` got a bit more work than the other
modules, but the code is "cleaner" now IMHO.

`which-function-in-header-line.el` still has problems because it
references my treepy-based mode line editing stuff that's only in
init.el, so byte compilation as above doesn't know that some of those
forms are macros.  I just deleted the resulting `.elc` file for it for
now (because otherwise persp-mode couldn't restore frames upon
restart... I think).