Turns out I like default scroll-conservatively
With 101 for scroll-conservatively, if I start a search from the top
of my init.el for anything (way?) off the screen, when I land on the
match Emacs doesn't recenter on the match.  That's annoying.  I think
50 might have both worked around the mentioned bug as well as letting
swiper recenter when jumping to a search match, but for now I'll just
stick with the default (0) and... move back to emacs-27 branch.