Rewrite company duplicate removal
I was foolishly sorting the candidates, and even doing so *after*
company-prescient sorted them for me, which I found very confusing and
frustrating.  New approach should preserve order, and will let me
change the priority of candidates among duplicates in the future,
should the need arise.

The code might not be beautiful.