Use buffer-chars-modified-tick instead of buffer-modified-tick
This was meant to suppress rebuilding the imenu index when the buffer
hadn't changed.  What I missed was that changes to text
properties (and what else?) increase `buffer-modified-tick`.  I
noticed this when scrolling through a very long Apache configuration
file was really slow.  Profiler pointed at imenu rebuilding as the
culprit.

This code now uses `buffer-chars-modified-tick` instead, to detect
when the content has actually changed.  If I find a case in the future
where text properties (or what else?) affects imenu index generation,
I will need to revisit this change.

An alternative to this could be a simple cap on "number of times per
second that imenu reindexing can run", maybe based on the size of the
file or how long the last reindexing took.