Split my utility functions out into their own file
This was actually so I could have one of my recipes import these
functions.  I've since rewritten the recipes not to need that
function (`my:with-eval-after-all-load`), but I decided to keep this
work as it's... well, it's done, at least.

Incidentally, I fixed an error involving the unbound symbol `null` in
one of the functions.  See if you can figure out which!  (I think it
was `my:add-to-list-before`---thanks to the byte compiler.)