Must run "terraform validate" from "root module"
The root module is defined as "wherever you run terraform from".
¯\_(ツ)_/¯

My best interpretation of this is: you needed to already have run
e.g. `terraform init` from the root module in order to install
providers and/or modules.  Without having done that, you can't expect
`terraform validate` to even work (i.e. my exact motivation for this
change).  Therefore we can search for this directory to determine the
root module.

This will not work in all situations.  Feel free to suggest me a
better way to find the root module.