Increase gc-cons-threshold after some testing
I was experiencing very slow Markdown editing when in a file with a
lot of `underscores_in_inline_code`.  This is apparently a known
problem in markdown-mode.  In the process of troubleshooting this,
though, I found that GC was taking a lot of time.  I followed (I think
it was) Eli's advice to try doubling the built-in `gc-cons-threshold`
until the time spent in GC while editing a Markdown file dropped to an
acceptably low percentage of CPU time.