Disable global-hl-line-mode in vterm-mode
BTW it turns out that `global-hl-line-mode` isn't a global variant of
a buffer-local mode, it's this other thing.  I think it was made this
way specifically so that it only highlights the line in the active
buffer, rather than the current line in every buffer.  The comments at
the top of `hl-line.el` give more details, including the authors
notice that "you probably don't want to use this".