The triumphant return of filladapt
In some brief testing, I think this will fix the problems I've had
with the built-in adaptive filling stuff not being smart enough to
pick out what is and isn't a prefix (e.g. bulleted lists, numbered
lists).  It's now part of GNU ELPA, so I hope this means it will be
maintained, FSVO "maintained".