Turn on display-fill-column-indicator-mode sometimes
I used to have the long lines visualization turned on for
whitespace-mode, but that got irritating and ugly.  fci-mode looks
nice now, at least on macOS--once you make modus-themes behave, which
was *really* hard and I'm pretty pissed about it right now.

Interesting that I seem to need to use `setq-default` here.  Is this
because the setting applied during init is somehow made "buffer-local"
to the buffer which is later killed?  I don't know.