Fix load order problems with gptel passwords
The macro to make the gptel backend skips it if it can't find the
password, function, I believe.  It didn't find the password function
because I had wrapped all the password getter functions in
init-local.el with `with-eval-after-load`, so I clearly lost the
"after load order lottery".

Bit jank to move this one macro out to its own file, but it works now.